+++ tests/async/threads.py
@@ -1,4 +1,4 @@
-from pulsar import ThreadPool, NOT_DONE
+from pulsar import ThreadPool, async_while, get_request_loop
 from pulsar.utils.pep import get_event_loop
 from pulsar.apps.test import unittest
 
@@ -21,12 +21,15 @@ class TestThreadQueue(unittest.TestCase):
         self.assertEqual(pool._state, 0)
         self.assertEqual(pool.status, 'running')
         self.assertEqual(pool.event_loop, get_event_loop())
-        #give a chanse to start the pool
-        yield NOT_DONE
+        self.assertNotEqual(pool.event_loop, get_request_loop())
+        #give a chance to start the pool
+        yield async_while(3, lambda: not pool.num_threads)
         self.assertEqual(pool.num_threads, 2)
         pool.close()
         self.assertEqual(pool._state, 1)
         self.assertEqual(pool.status, 'closed')
         pool.join()
+        yield async_while(3, lambda: pool.num_threads)
+        self.assertFalse(pool.num_threads)