+++ CHANGELOG.rst
@@ -3,11 +3,13 @@ Ver. 0.6.0 - Development
 * Added :ref:`--io option <setting-poller>` for controlling the default
   :class:`pulsar.Poller`.
 * Added support of ``kqueue`` :class:`pulsar.Poller` for systems supporting it.
+* New asynchronous :class:`pulsar.ThreadPool` for CPU bound operations.
 * Modified :class:`pulsar.Failure` implementation to handle one ``exc_info`` only and
   better handling of unlogged failures.
 * Critical bug fix in python 3 WSGI server.
 * Added an asynchronous FIFO :class:`pulsar.Queue`.
 * Proxy server example uses the new :class:`pulsar.Queue`.
+* Added :func:`pulsar.async_while` utility function.
 * Socket servers handle IPV6 addresses.
 * Added :ref:`SSL support <socket-server-ssl>` for socket servers.
 * :ref:`Actor's hooks can be asynchronous <actor-hooks>`.